extract re-usable logic into base class for alternate sub-class for different http clients
