+++ test/solr_json_writer_test.rb
@@ -256,9 +256,9 @@ describe "Traject::SolrJsonWriter" do
       @writer.put context_with({"id" => "one", "key" => ["value1", "value2"]})
       @writer.close
 
-      last_solr_get = @fake_http_client.get_args.last
+      last_solr_post = @fake_http_client.post_args.last
 
-      assert_equal "http://example.com/update/json?commit=true", last_solr_get[0]
+      assert_equal "http://example.com/update/json?commit=true", last_solr_post[0]
     end
 
     it "commits on close with commit_solr_update_args" do
@@ -270,24 +270,24 @@ describe "Traject::SolrJsonWriter" do
       @writer.put context_with({"id" => "one", "key" => ["value1", "value2"]})
       @writer.close
 
-      last_solr_get = @fake_http_client.get_args.last
+      last_solr_post = @fake_http_client.post_args.last
 
-      assert_equal "http://example.com/update/json?softCommit=true", last_solr_get[0]
+      assert_equal "http://example.com/update/json?softCommit=true", last_solr_post[0]
     end
 
     it "can manually send commit" do
       @writer = create_writer("solr.url" => "http://example.com")
       @writer.commit
 
-      last_solr_get = @fake_http_client.get_args.last
-      assert_equal "http://example.com/update/json?commit=true", last_solr_get[0]
+      last_solr_post = @fake_http_client.post_args.last
+      assert_equal "http://example.com/update/json?commit=true", last_solr_post[0]
     end
 
     it "can manually send commit with specified args" do
       @writer = create_writer("solr.url" => "http://example.com", "solr_writer.commit_solr_update_args" => { softCommit: true })
       @writer.commit(commit: true, optimize: true, waitFlush: false)
-      last_solr_get = @fake_http_client.get_args.last
-      assert_equal "http://example.com/update/json?commit=true&optimize=true&waitFlush=false", last_solr_get[0]
+      last_solr_post = @fake_http_client.post_args.last
+      assert_equal "http://example.com/update/json?commit=true&optimize=true&waitFlush=false", last_solr_post[0]
     end
 
     it "uses commit_solr_update_args settings by default" do
@@ -297,8 +297,8 @@ describe "Traject::SolrJsonWriter" do
       )
       @writer.commit
 
-      last_solr_get = @fake_http_client.get_args.last
-      assert_equal "http://example.com/update/json?softCommit=true", last_solr_get[0]
+      last_solr_post = @fake_http_client.post_args.last
+      assert_equal "http://example.com/update/json?softCommit=true", last_solr_post[0]
     end
 
     it "overrides commit_solr_update_args with method arg" do
@@ -308,8 +308,8 @@ describe "Traject::SolrJsonWriter" do
       )
       @writer.commit(commit: true)
 
-      last_solr_get = @fake_http_client.get_args.last
-      assert_equal "http://example.com/update/json?commit=true", last_solr_get[0]
+      last_solr_post = @fake_http_client.post_args.last
+      assert_equal "http://example.com/update/json?commit=true", last_solr_post[0]
     end
   end
